+++ Makefile
@@ -1,4 +1,4 @@
-# Makefile for sz-sdk-nodejs-grpc
+# Makefile for sz-sdk-typescript-grpc
 
 # -----------------------------------------------------------------------------
 # Variables
@@ -12,7 +12,7 @@ MAKEFILE_PATH := $(abspath $(firstword $(MAKEFILE_LIST)))
 MAKEFILE_DIRECTORY := $(shell dirname $(MAKEFILE_PATH))
 DIST_DIRECTORY := $(MAKEFILE_DIRECTORY)/dist
 NS_DIRECTORY := $(DIST_DIRECTORY)/@senzing
-TARGET_DIRECTORY := $(DIST_DIRECTORY)/@senzing/sz-sdk-nodejs-grpc
+TARGET_DIRECTORY := $(DIST_DIRECTORY)/@senzing/sz-sdk-typescript-grpc
 BUILD_VERSION := $(shell git describe --always --tags --abbrev=0 --dirty  | sed 's/v//')
 BUILD_TAG := $(shell git describe --always --tags --abbrev=0  | sed 's/v//')
 BUILD_ITERATION := $(shell git log $(BUILD_TAG)..HEAD --oneline | wc -l | sed 's/^ *//')
@@ -77,7 +77,7 @@ documentation:
 .PHONY: clean
 clean:
 	@rm -rf $(TARGET_DIRECTORY)/* || true
-	@rm -rf $(NS_DIRECTORY)/senzing-sz-sdk-nodejs-grpc*.tgz || true
+	@rm -rf $(NS_DIRECTORY)/senzing-sz-sdk-typescript-grpc*.tgz || true
 	@rm -rf $(NS_DIRECTORY)/*.tsbuildinfo || true
 
 # -----------------------------------------------------------------------------
