+++ examples/szconfig/getDataSources.ts
@@ -1,4 +1,4 @@
-import { SzAbstractFactory, SzAbstractFactoryOptions } from '../../dist/@senzing/sz-sdk-nodejs-grpc';
+import { SzAbstractFactory, SzAbstractFactoryOptions } from '../../dist/@senzing/sz-sdk-typescript-grpc';
 
 const szParamFactory = new SzAbstractFactory(`0.0.0.0:8261`);
 const szConfig      = szParamFactory.createConfig();
