+++ examples/szproduct/getVersion.js
@@ -1,10 +1,12 @@
-import { SzAbstractFactory } from '../../dist/@senzing/sz-sdk-nodejs-grpc';
-const szParamFactory = new SzAbstractFactory(`0.0.0.0:8261`);
-const szProduct = szParamFactory.createProduct();
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+var sz_sdk_typescript_grpc_1 = require("../../dist/@senzing/sz-sdk-typescript-grpc");
+var szParamFactory = new sz_sdk_typescript_grpc_1.SzAbstractFactory("0.0.0.0:8261");
+var szProduct = szParamFactory.createProduct();
 szProduct.getVersion().
-    then((result) => {
+    then(function (result) {
     console.log("RESPONSE:\n\r", result);
 }).
-    catch((err) => {
+    catch(function (err) {
     console.error(err);
 });
