+++ package.json
@@ -3,9 +3,9 @@
   "version": "0.0.1",
   "scripts": {
     "build": "npm run tsc && node --disable-warning=ExperimentalWarning build.js",
-    "clean": "rm -fR dist/@senzing/sz-sdk-nodejs-grpc && rm -fR dist/@senzing/sz-sdk-nodejs-grpc.tsbuildinfo",
+    "clean": "rm -fR dist/@senzing/sz-sdk-typescript-grpc && rm -fR dist/@senzing/sz-sdk-typescript-grpc.tsbuildinfo",
     "compile": "npm run tsc",
-    "package": "cd dist/\\@senzing\\/sz-sdk-nodejs-grpc && npm pack",
+    "package": "cd dist/\\@senzing\\/sz-sdk-typescript-grpc && npm pack",
     "postpackage": "node post-pack.js",
     "examples:ts:helloworld": "tsx examples/static/helloworld.ts",
     "test": "echo \"Error: no test specified\" && exit 1",
