+++ post-pack.js
@@ -24,4 +24,4 @@ function movePackage(sourcePattern, destination) {
     });
 }
 
-movePackage(path.join('dist','@senzing','sz-sdk-nodejs-grpc','*.tgz'), path.join('dist','@senzing'));
+movePackage(path.join('dist','@senzing','sz-sdk-typescript-grpc','*.tgz'), path.join('dist','@senzing'));
